refactor: ConditionFilter 클래스 분리
