+++ backend/src/main/java/com/woowacourse/momo/group/repository/GroupRepositoryCustomImpl.java
@@ -3,10 +3,8 @@ package com.woowacourse.momo.group.repository;
 import static com.woowacourse.momo.group.domain.QGroup.group;
 import static com.woowacourse.momo.group.domain.participant.QParticipant.participant;
 
-import java.time.LocalDateTime;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Optional;
 import java.util.function.Supplier;
 
 import javax.persistence.EntityManager;
@@ -16,17 +14,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.support.PageableExecutionUtils;
 import org.springframework.stereotype.Repository;
 
-import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
-import com.querydsl.core.types.dsl.Expressions;
-import com.querydsl.core.types.dsl.NumberExpression;
-import com.querydsl.core.types.dsl.NumberPath;
 import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.JPQLQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
-import com.woowacourse.momo.category.domain.Category;
 import com.woowacourse.momo.group.domain.Group;
 import com.woowacourse.momo.group.domain.participant.Participant;
 import com.woowacourse.momo.member.domain.Member;
@@ -35,9 +28,11 @@ import com.woowacourse.momo.member.domain.Member;
 public class GroupRepositoryCustomImpl implements GroupRepositoryCustom {
 
     private final JPAQueryFactory queryFactory;
+    private final ConditionFilter conditionFilter;
 
-    public GroupRepositoryCustomImpl(EntityManager em) {
-        this.queryFactory = new JPAQueryFactory(em);
+    public GroupRepositoryCustomImpl(EntityManager entityManager) {
+        this.queryFactory = new JPAQueryFactory(entityManager);
+        this.conditionFilter = new ConditionFilter();
     }
 
     @Override
@@ -60,7 +55,7 @@ public class GroupRepositoryCustomImpl implements GroupRepositoryCustom {
                 .selectFrom(group)
                 .where(
                         supplier.get(),
-                        filterByCondition(condition)
+                        conditionFilter.filterByCondition(condition)
                 )
                 .orderBy(orderByDeadlineAsc(condition.orderByDeadline()).toArray(OrderSpecifier[]::new))
                 .offset(pageable.getOffset())
@@ -92,63 +87,6 @@ public class GroupRepositoryCustomImpl implements GroupRepositoryCustom {
                 .where(participant.member.eq(member));
 
         return group.participants.participants.contains(targetMember);
-    }
-
-    private BooleanBuilder filterByCondition(FindCondition condition) {
-        BooleanBuilder booleanBuilder = new BooleanBuilder();
-
-        excludeFinished(booleanBuilder, condition.excludeFinished());
-        filterByCategory(booleanBuilder, condition.getCategory());
-        containKeyword(booleanBuilder, condition.getKeyword());
-        afterNow(booleanBuilder, condition.orderByDeadline());
-
-        return booleanBuilder;
-    }
-
-    private void excludeFinished(BooleanBuilder booleanBuilder, boolean condition) {
-        if (condition) {
-            booleanBuilder.and(afterNow()
-                    .and(notClosedEarly())
-                    .and(isNotParticipantsFull()));
-        }
-    }
-
-    private void filterByCategory(BooleanBuilder booleanBuilder, Optional<Long> condition) {
-        condition.ifPresent(categoryId -> {
-            Category category = Category.from(categoryId);
-            booleanBuilder.and(group.category.eq(category));
-        });
-    }
-
-    private void containKeyword(BooleanBuilder booleanBuilder, Optional<String> condition) {
-        condition.ifPresent(keyword -> {
-            BooleanExpression nameContains = group.name.value.contains(keyword);
-            BooleanExpression descriptionContains = group.description.contains(keyword);
-
-            booleanBuilder.and(nameContains.or(descriptionContains));
-        });
-    }
-
-    private void afterNow(BooleanBuilder booleanBuilder, boolean condition) {
-        if (condition) {
-            booleanBuilder.and(afterNow());
-        }
-    }
-
-    private BooleanExpression isNotParticipantsFull() {
-        NumberPath<Integer> capacity = group.participants.capacity.value;
-        NumberExpression<Integer> participantsSize = group.participants.participants.size()
-                .add(Expressions.constant(1));
-
-        return capacity.gt(participantsSize);
-    }
-
-    private BooleanExpression notClosedEarly() {
-        return group.closedEarly.isFalse();
-    }
-
-    private BooleanExpression afterNow() {
-        return group.calendar.deadline.value.gt(LocalDateTime.now());
     }
 
     private List<OrderSpecifier<?>> orderByDeadlineAsc(boolean orderByDeadline) {
